Create workspace for vite plugins
This adds a new npm workspace for the plugins & build utils used
throughout the frontend codebase. This allows deduplicating some helpers
(reading replica host, canister id) and move some concerns (building)
out of the main workspace.
