+++ demos/vc_issuer/package.json
@@ -11,11 +11,11 @@
     "@dfinity/principal": "^0.19.3",
     "@dfinity/utils": "^0.0.20",
     "buffer": "^6.0.3",
+    "html-minifier-terser": "^7.2.0",
+    "http-proxy": "^1.18.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "http-proxy": "^1.18.1",
-    "vite-plugin-compression": "^0.5.1",
-    "html-minifier-terser": "^7.2.0"
+    "vite-plugin-compression": "^0.5.1"
   },
   "scripts": {
     "dev": "vite --config ./vite.config.ts",
@@ -27,12 +27,13 @@
     "generate:js": "didc bind ./vc_demo_issuer.did -t js > ./app/generated/vc_issuer_idl.js"
   },
   "devDependencies": {
+    "@dfinity/internet-identity-vite-plugins": "*",
     "@types/html-minifier-terser": "^7.0.0",
     "@types/http-proxy": "^1.17.14",
     "@types/react": "^18.2.38",
     "@types/react-dom": "^18.2.17",
     "@vitejs/plugin-react": "^4.2.0",
-    "vite": "^4.5.1",
-    "typescript": "5.2.2"
+    "typescript": "5.2.2",
+    "vite": "^4.5.1"
   }
 }
